Relocate BatchContainer, and experiment with ToOwned
