+++ src/trace/layers/mod.rs
@@ -4,9 +4,6 @@
 //! in the next layer. Similarly, ranges of elements in the layer itself may correspond
 //! to single elements in the layer above.
 
-use timely::container::columnation::TimelyStack;
-use timely::container::columnation::Columnation;
-
 pub mod ordered;
 pub mod ordered_leaf;
 // pub mod hashed;
@@ -107,152 +104,4 @@ pub trait Cursor<Storage> {
     fn rewind(&mut self, storage: &Storage);
     /// Repositions the cursor to a different range of values.
     fn reposition(&mut self, storage: &Storage, lower: usize, upper: usize);
-}
-
-/// A general-purpose container resembling `Vec<T>`.
-pub trait BatchContainer: Default {
-    /// The type of contained item.
-    type Item;
-    /// Inserts an owned item.
-    fn push(&mut self, item: Self::Item);
-    /// Inserts a borrowed item.
-    fn copy(&mut self, item: &Self::Item);
-    /// Extends from a slice of items.
-    fn copy_slice(&mut self, slice: &[Self::Item]);
-    /// Extends from a range of items in another`Self`.
-    fn copy_range(&mut self, other: &Self, start: usize, end: usize);
-    /// Creates a new container with sufficient capacity.
-    fn with_capacity(size: usize) -> Self;
-    /// Reserves additional capacity.
-    fn reserve(&mut self, additional: usize);
-    /// Creates a new container with sufficient capacity.
-    fn merge_capacity(cont1: &Self, cont2: &Self) -> Self;
-
-    /// Reference to the element at this position.
-    fn index(&self, index: usize) -> &Self::Item;
-    /// Number of contained elements
-    fn len(&self) -> usize;
-    /// Returns the last item if the container is non-empty.
-    fn last(&self) -> Option<&Self::Item> {
-        if self.len() > 0 {
-            Some(self.index(self.len()-1))
-        }
-        else {
-            None
-        }
-    }
-    
-    /// Reports the number of elements satisfing the predicate.
-    ///
-    /// This methods *relies strongly* on the assumption that the predicate
-    /// stays false once it becomes false, a joint property of the predicate
-    /// and the layout of `Self. This allows `advance` to use exponential search to
-    /// count the number of elements in time logarithmic in the result.
-    fn advance<F: Fn(&Self::Item)->bool>(&self, start: usize, end: usize, function: F) -> usize {
-
-        let small_limit = 8;
-
-        // Exponential seach if the answer isn't within `small_limit`.
-        if end > start + small_limit && function(self.index(start + small_limit)) {
-
-            // start with no advance
-            let mut index = small_limit + 1;
-            if start + index < end && function(self.index(start + index)) {
-
-                // advance in exponentially growing steps.
-                let mut step = 1;
-                while start + index + step < end && function(self.index(start + index + step)) {
-                    index += step;
-                    step = step << 1;
-                }
-
-                // advance in exponentially shrinking steps.
-                step = step >> 1;
-                while step > 0 {
-                    if start + index + step < end && function(self.index(start + index + step)) {
-                        index += step;
-                    }
-                    step = step >> 1;
-                }
-
-                index += 1;
-            }
-
-            index
-        }
-        else {
-            let limit = std::cmp::min(end, start + small_limit);
-            (start .. limit).filter(|x| function(self.index(*x))).count()
-        }
-    }
-}
-
-impl<T: Clone> BatchContainer for Vec<T> {
-    type Item = T;
-    fn push(&mut self, item: T) {
-        self.push(item);
-    }
-    fn copy(&mut self, item: &T) {
-        self.push(item.clone());
-    }
-    fn copy_slice(&mut self, slice: &[T]) {
-        self.extend_from_slice(slice);
-    }
-    fn copy_range(&mut self, other: &Self, start: usize, end: usize) {
-        self.extend_from_slice(&other[start .. end]);
-    }
-    fn with_capacity(size: usize) -> Self {
-        Vec::with_capacity(size)
-    }
-    fn reserve(&mut self, additional: usize) {
-        self.reserve(additional);
-    }
-    fn merge_capacity(cont1: &Self, cont2: &Self) -> Self {
-        Vec::with_capacity(cont1.len() + cont2.len())
-    }
-    fn index(&self, index: usize) -> &Self::Item {
-        &self[index]
-    }
-    fn len(&self) -> usize {
-        self[..].len()
-    }
-}
-
-impl<T: Columnation> BatchContainer for TimelyStack<T> {
-    type Item = T;
-    fn push(&mut self, item: T) {
-        self.copy(&item);
-    }
-    fn copy(&mut self, item: &T) {
-        self.copy(item);
-    }
-    fn copy_slice(&mut self, slice: &[T]) {
-        self.reserve_items(slice.iter());
-        for item in slice.iter() {
-            self.copy(item);
-        }
-    }
-    fn copy_range(&mut self, other: &Self, start: usize, end: usize) {
-        let slice = &other[start .. end];
-        self.reserve_items(slice.iter());
-        for item in slice.iter() {
-            self.copy(item);
-        }
-    }
-    fn with_capacity(size: usize) -> Self {
-        Self::with_capacity(size)
-    }
-    fn reserve(&mut self, _additional: usize) {
-    }
-    fn merge_capacity(cont1: &Self, cont2: &Self) -> Self {
-        let mut new = Self::default();
-        new.reserve_regions(std::iter::once(cont1).chain(std::iter::once(cont2)));
-        new
-    }
-    fn index(&self, index: usize) -> &Self::Item {
-        &self[index]
-    }
-    fn len(&self) -> usize {
-        self[..].len()
-    }
 }
